Calculate Pairwise Distance Matrix

Calculate Pairwise Distance Matrix

Usage

calculate_pairwise_distances_matrix(
  data,
  x_col = "x",
  y_col = "y",
  distance_fun = distGeo
)

Arguments

data

A data frame containing site coordinates and IDs.

x_col

Name of the longitude column in data (default: "x").

y_col

Name of the latitude column in data (default: "y").

distance_fun

The distance function to use (default: distGeo).

Value

A data frame containing pairwise distances between sites (in km).

Examples

# default usage when your coords are in columns "x" and "y"
dist_df_default <- calculate_pairwise_distances_matrix(data = my_sites_df)
#> Error: object 'my_sites_df' not found
head(dist_df_default)
#> Error: object 'dist_df_default' not found

# specify custom column names for longitude and latitude
dist_df <- calculate_pairwise_distances_matrix(
  data  = my_sites_df,
  x_col = "centroid_lon",
  y_col = "centroid_lat"
)
#> Error: object 'my_sites_df' not found
head(dist_df)
#> Error: object 'dist_df' not found