Predict Dissimilarity Using Environmental Data

This function predicts turnover or zeta diversity values using models from zetadiv::Zeta.msgdm and new environmental data.

Usage

predict_dissim(
  block_sp,
  sbe_scaled,
  zeta_model,
  mean_rich,
  mean_turn,
  sbs_xy,
  rsa
)

Arguments

block_sp

Data frame containing geographic and environmental data.

sbe_scaled

Scaled environmental predictors.

zeta_model

Model object from zetadiv::Zeta.msgdm.

mean_rich

Data frame with mean richness values for calibration.

mean_turn

Data frame with mean turnover values for calibration.

sbs_xy

Data frame with spatial coordinates.

rsa

Spatial object for overlaying boundaries on plots.

Value

A data frame with predicted dissimilarity and other metrics.

Examples

# Example usage:
result <- predict_dissim(block_sp, sbe_scaled, zeta_model, mean_rich, mean_turn, sbs_xy, rsa)
#> Loading required package: scam
#> Loading required package: mgcv
#> Loading required package: nlme
#> 
#> Attaching package: ‘nlme’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     collapse
#> This is mgcv 1.9-1. For overview type 'help("mgcv-package")'.
#> This is scam 1.2-18.
#> This is zetadiv 1.2.1
#> Package "zetadiv" was built under R.4.1.3
#> Error: object 'block_sp' not found
head(result)
#>                                                                        
#> 1 function (future, ...)                                               
#> 2 {                                                                    
#> 3     if (inherits(future, "Future") && inherits(future[[".journal"]], 
#> 4         "FutureJournal")) {                                          
#> 5         start <- Sys.time()                                          
#> 6         on.exit({